extract run support bundle route
